requirement.txt
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
cmd2==2.4.3
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
EbookLib==0.18
exceptiongroup==1.2.1
executing==2.0.1
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
loguru==0.7.2
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
platformdirs==4.2.2
PrettyPrintTree==2.0.0
prompt_toolkit==3.0.45
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
pyperclip==1.8.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.3
requests==2.32.3
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13
win32-setctime==1.1.0
zipp==3.19.1
grequests